Update ThreadPool class
Rewrite the internal representation of the ThreadPool class. The class will store simple functions in the work queue that wraps a heap allocated packed_task which will execute the required function. This will allow the ThreadPool to return a std::future that has a valid state and can be used to retrieve the task result once it is ready.