Avoid heap allocation of event objects
Change the way the event objects were allocated to avoid extra heap memory costs that are not needed for the associated smart pointer. Actually, class Event is a very simple construct -especially after moving the random number generator out of it- and there is no concern about keeping it on stack and allow it to be copyable. This enhancment further reduced the space complexity for running 1million+ events in a simulation