Move the ownership of the random number generator from a per module to a per thread ownership
Changing the ownership of the random number generator from the module class itself resolves the need for locking and uses synchronization to access/use the generator. The idea is simple, let every thread has its own random generator with a storage specification thread_local. Then this generator will be passed down to the module to use it for its own purposes as an argument after seeding it using the seed encapsulated in the event object. Now every thread, event and module is completely independent to allow better scalling and minimze the overhead added by using multithreaded execution